Feature/eint23 446 instruct loan proposal

See merge request intellecteu/blockchain/clients/equilend-integration/eint-1source-integration!71

+++ src/main/java/com/intellecteu/onesource/integration/model/onesource/TradeAgreement.java
@@ -36,7 +36,7 @@ public class TradeAgreement implements Reconcilable {
     private Rate rate;
     private Integer quantity;
     private CurrencyCd billingCurrency;
-    private Integer dividendRatePct;
+    private Double dividendRatePct;
     private LocalDate tradeDate;
     private TermType termType;
     private LocalDate termDate;
